+++ plugins/network/__init__.py
@@ -20,193 +20,98 @@
 #########################################################################
 
 import logging
-import asynchat
-import asyncore
-import socket
 import threading
+import socket
 import urllib.request
 import urllib.parse
 import urllib.error
+import lib.connection
 
 logger = logging.getLogger('')
 
 
-class TCPHandler(asynchat.async_chat):
+class TCPHandler(lib.connection.Connection):
 
-    terminator = '\n'.encode()
-
-    def __init__(self, socket_map, parser, dest, sock, source):
-        asynchat.async_chat.__init__(self, sock=sock, map=socket_map)
+    def __init__(self, parser, dest, sock, source):
+        lib.connection.Connection.__init__(self, sock, source)
+        self.terminator = b'\n'
         self.parser = parser
         self._lock = threading.Lock()
         self.dest = dest
-        self.buffer = bytearray()
         self.source = source
 
-    def initiate_send(self):
-        self._lock.acquire()
-        asynchat.async_chat.initiate_send(self)
-        self._lock.release()
-
-    def collect_incoming_data(self, data):
-        self.buffer.extend(data)
-
-    def found_terminator(self):
-        data = self.buffer
-        self.buffer = bytearray()
+    def found_terminator(self, data):
         self.parser(self.source, self.dest, data.strip())
-        try:
-            self.shutdown(socket.SHUT_RDWR)
-        except:
-            pass
-        try:
-            self.close()
-        except:
-            pass
+        self.close()
 
 
-class TCPDispatcher(asyncore.dispatcher):
+class TCPDispatcher(lib.connection.Server):
 
-    def __init__(self, parser, socket_map, ip, port):
-        asyncore.dispatcher.__init__(self, map=socket_map)
-        self.socket_map = socket_map
+    def __init__(self, parser, ip, port):
+        lib.connection.Server.__init__(self, ip, port)
         self.parser = parser
         self.dest = 'tcp:' + ip + ':' + port
-        try:
-            self.create_socket(socket.AF_INET, socket.SOCK_STREAM)
-            self.set_reuse_addr()
-            self.bind((ip, int(port)))
-            self.listen(5)
-            self.listening = True
-        except Exception:
-            logger.error("Could not bind TCP socket on {}:{}".format(ip, port))
-            self.listening = False
+        self.connect()
 
-    def handle_accept(self):
-        pair = self.accept()
-        if pair is not None:
-            sock, (ip, port) = pair
-            logger.debug('{} Incoming connection from {}:{}'.format(self.dest, ip, port))
-            TCPHandler(self.socket_map, self.parser, self.dest, sock, ip)
+    def handle_connection(self):
+        sock, address = self.accept()
+        if sock is None:
+            return
+        TCPHandler(self.parser, self.dest, sock, address)
 
 
-class HTTPHandler(asynchat.async_chat):
+class HTTPHandler(lib.connection.Connection):
 
-    terminator = "\r\n\r\n".encode()
-
-    def __init__(self, socket_map, parser, dest, sock, source):
-        asynchat.async_chat.__init__(self, sock=sock, map=socket_map)
+    def __init__(self, parser, dest, sock, source):
+        lib.connection.Connection.__init__(self, sock, source)
+        self.terminator = b"\r\n\r\n"
         self.parser = parser
         self._lock = threading.Lock()
         self.dest = dest
-        self.buffer = bytearray()
         self.source = source
 
-    def initiate_send(self):
-        self._lock.acquire()
-        asynchat.async_chat.initiate_send(self)
-        self._lock.release()
-
-    def collect_incoming_data(self, data):
-        self.buffer.extend(data)
-
-    def found_terminator(self):
-        data = self.buffer
-        self.buffer = bytearray()
+    def found_terminator(self, data):
         for line in data.decode().splitlines():
             if line.startswith('GET'):
                 request = line.split(' ')[1].strip('/')
                 self.parser(self.source, self.dest, urllib.parse.unquote(request))
                 break
-        try:
-            self.shutdown(socket.SHUT_RDWR)
-        except:
-            pass
-        try:
-            self.close()
-        except:
-            pass
+        self.close()
 
 
-class HTTPDispatcher(asyncore.dispatcher):
+class HTTPDispatcher(lib.connection.Server):
 
-    def __init__(self, parser, socket_map, ip, port):
-        asyncore.dispatcher.__init__(self, map=socket_map)
-        self.socket_map = socket_map
+    def __init__(self, parser, ip, port):
+        lib.connection.Server.__init__(self, ip, port)
         self.parser = parser
         self.dest = 'http:' + ip + ':' + port
-        try:
-            self.create_socket(socket.AF_INET, socket.SOCK_STREAM)
-            self.set_reuse_addr()
-            self.bind((ip, int(port)))
-            self.listen(5)
-            self.listening = True
-        except Exception:
-            logger.error("Could not bind TCP socket for HTTP on {}:{}".format(ip, port))
-            self.listening = False
+        self.connect()
 
-    def handle_accept(self):
-        pair = self.accept()
-        if pair is not None:
-            sock, (ip, port) = pair
-            logger.debug('{} Incoming connection from {}:{}'.format(self.dest, ip, port))
-            HTTPHandler(self.socket_map, self.parser, self.dest, sock, ip)
+    def handle_connection(self):
+        sock, address = self.accept()
+        if sock is None:
+            return
+        HTTPHandler(self.parser, self.dest, sock, address)
 
 
-class UDPDispatcher(asyncore.dispatcher):
+class UDPDispatcher(lib.connection.Server):
 
-    def __init__(self, parser, socket_map, ip, port):
-        asyncore.dispatcher.__init__(self, map=socket_map)
+    def __init__(self, parser, ip, port):
+        lib.connection.Server.__init__(self, ip, port, proto='UDP')
         self.dest = 'udp:' + ip + ':' + port
         self.parser = parser
+        self.connect()
+
+    def handle_connection(self):
         try:
-            self.create_socket(socket.AF_INET, socket.SOCK_DGRAM)
-            self.set_reuse_addr()
-            self.bind((ip, int(port)))
-            self.listening = True
-        except Exception:
-            logger.error("Could not bind UDP socket on {}:{}".format(ip, port))
-            self.listening = False
-
-    def handle_read(self):
-        data, (ip, port) = self.recvfrom(4096)
-        logger.debug('{} Incoming connection from {}:{}'.format(self.dest, ip, port))
-        self.parser(ip, self.dest, data.strip())
-
-    def writable(self):
-        return False
-
-
-class UDPSend(asyncore.dispatcher_with_send):
-
-    def __init__(self, socket_map, host, port, data):
-        asyncore.dispatcher_with_send.__init__(self, map=socket_map)
-        try:
-            self.create_socket(socket.AF_INET, socket.SOCK_DGRAM)
-            self.connect((host, int(port)))
-        except Exception:
-            logger.warning("Could not connect to {}:{}, to send data: {}.".format(host, port, data))
+            data, addr = self._socket.recvfrom(4096)
+            ip = addr[0]
+            addr = "{}:{}".format(addr[0], addr[1])
+            logger.debug("{}: incoming connection from {} to {}".format(self._name, addr, self.address))
+        except Exception as e:
+            logger.exception("{}: {}".format(self._name, e))
             return
-        self.send(data)
-
-    def writable(self):
-        if len(self.out_buffer) > 0 and self.connected:
-            return True
-        else:
-            self.close()
-            return False
-
-    def readable(self):
-        return False
-
-    def handle_connect(self):
-        pass
-
-    def handle_close(self):
-        try:
-            self.close()
-        except:
-            pass
+        self.parser(ip, self.dest, data.decode().strip())
 
 
 class Network():
@@ -219,7 +124,6 @@ class Network():
 
     def __init__(self, smarthome, ip='0.0.0.0', port='2727', udp='no', tcp='no', http='no', udp_acl='*', tcp_acl='*', http_acl='*'):
         self._sh = smarthome
-        self.socket_map = smarthome.socket_map
         self.tcp_acl = self.parse_acl(tcp_acl)
         self.udp_acl = self.parse_acl(udp_acl)
         self.http_acl = self.parse_acl(http_acl)
@@ -231,20 +135,30 @@ class Network():
             self.add_listener('http', ip, http, http_acl, generic=True)
 
     def udp(self, host, port, data):
-        UDPSend(self.socket_map, host, port, data)
+        try:
+            family, type, proto, canonname, sockaddr = socket.getaddrinfo(host, port)[0]
+            sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+            sock.sendto(data.encode(), (sockaddr[0], sockaddr[1]))
+            sock.close()
+            del(sock)
+        except Exception as e:
+            logger.warning("UDP: Problem sending data to {}:{}: ".format(host, port, e))
+            pass
+        else:
+            logger.debug("UDP: Sending data to {}:{}: ".format(host, port, data))
 
     def add_listener(self, proto, ip, port, acl='*', generic=False):
         dest = proto + ':' + ip + ':' + port
         logger.debug("Adding listener on: {}".format(dest))
         if proto == 'tcp':
-            dispatcher = TCPDispatcher(self.parse_input, self.socket_map, ip, port)
+            dispatcher = TCPDispatcher(self.parse_input, ip, port)
         elif proto == 'udp':
-            dispatcher = UDPDispatcher(self.parse_input, self.socket_map, ip, port)
+            dispatcher = UDPDispatcher(self.parse_input, ip, port)
         elif proto == 'http':
-            dispatcher = HTTPDispatcher(self.parse_input, self.socket_map, ip, port)
+            dispatcher = HTTPDispatcher(self.parse_input, ip, port)
         else:
             return
-        if not dispatcher.listening:
+        if not dispatcher.connected:
             return False
         acl = self.parse_acl(acl)
         if generic:
@@ -364,6 +278,18 @@ class Network():
 
     def parse_item(self, item):
         self.parse_obj(item, 'item')
+        if 'nw_udp_send' in item.conf:
+            return self.update_item
+
+    def update_item(self, item, caller=None, source=None, dest=None):
+        if 'nw_udp_send' in item.conf:
+            addr, __, message = item.conf['nw_udp_send'].partition('=')
+            if message is None:
+                message = str(item()).encode()
+            else:
+                message = message.replace('itemvalue', str(item())).encode()
+            host, __, port = addr.partition(':')
+            self.udp(host, port, message)
 
     def parse_obj(self, obj, obj_type):
         # nw_acl, nw_udp, nw_tcp